requirements.txt
numpy
pandas
scipy
numba
threadpoolctl
scikit-learn
emoji
nltk
spacy
pattern
gensim
mip
langdetect
plotly.express
datetime
matplotlib
seaborn
openpyxl
